Create IAM 3
In this project, I worked on IAM Roles in AWS. The goal was to provide secure and limited access to specific AWS resources using roles. I created a role that gives full access to VPC and DynamoDB, but only allows two users (user1 and user2) to assume that role. After setting up the role and trust relationship, I tested it by logging in as user1 and successfully accessing the allowed services.

This helps in following the principle of least privilege and better access management.